@@ -5,9 +5,8 @@
     "description": "A VS Code plugin that visualizes Swagger/OpenAPI definitions",
     "version": "1.0.2",
     "engines": {
-        "vscode": "^1.70.0",
-        "node": "^18.13.0",
-        "npm": "^8.19.3"
+        "vscode": "^1.74.0",
+        "node": ">=20"
     },
     "license": "SEE LICENSE IN LICENSE.md",
     "categories": [
@@ -71,20 +70,20 @@
         }
     },
     "scripts": {
-        "vscode:prepublish": "npm run compile",
+        "vscode:prepublish": "yarn compile",
         "precompile": "rimraf dist && rimraf coverage",
         "compile": "tsc -p ./",
-        "build": "npm run compile",
+        "build": "yarn compile",
         "watch": "tsc -watch -p ./",
-        "lint": "eslint src --ext ts --fix",
-        "lint:check": "eslint src --ext ts",
-        "format": "npx prettier --write .",
-        "format:check": "npx prettier --check .",
-        "pretest": "npm run compile && npm run lint",
-        "test": "npx jest --maxWorkers=4 --logHeapUsage --coverage --coverageThreshold '{}' --forceExit --detectOpenHandles",
+        "lint": "eslint . --ext .ts",
+        "lint:check": "eslint . --ext .ts --max-warnings 0",
+        "format": "prettier --write .",
+        "format:check": "prettier --check .",
+        "pretest": "yarn compile && yarn lint",
+        "test": "jest --maxWorkers=4 --logHeapUsage --coverage --coverageThreshold '{}' --forceExit --detectOpenHandles",
         "prepare": "husky install",
-        "package": "npx vsce package",
-        "bumpup": "npm version patch --force"
+        "package": "vsce package",
+        "bumpup": "yarn version patch --no-git-tag-version"
     },
     "lint-staged": {
         "*.ts|.json": [
@@ -96,38 +95,39 @@
         ]
     },
     "devDependencies": {
-        "@commitlint/cli": "^17.6.3",
-        "@commitlint/config-conventional": "^17.6.3",
-        "@types/express": "^4.17.17",
-        "@types/jest": "^29.5.1",
+        "@commitlint/cli": "^19.8.1",
+        "@commitlint/config-conventional": "^19.8.1",
+        "@types/express": "^5.0.3",
+        "@types/jest": "^30.0.0",
         "@types/js-yaml": "^4.0.5",
         "@types/lodash": "^4.14.194",
-        "@types/node": "16.x",
-        "@types/vscode": "^1.70.0",
-        "@typescript-eslint/eslint-plugin": "^5.59.1",
-        "@typescript-eslint/parser": "^5.59.1",
+        "@types/node": "^24.0.10",
+        "@types/vscode": "^1.74.0",
+        "@typescript-eslint/eslint-plugin": "^8.35.1",
+        "@typescript-eslint/parser": "^8.35.1",
+        "typescript": "~5.3.3",
         "@vscode/test-electron": "^2.3.0",
-        "eslint": "^8.39.0",
+        "eslint": "^9.30.1",
         "eslint-plugin-filenames": "^1.3.2",
-        "husky": "^8.0.3",
-        "jest": "^29.5.0",
-        "lint-staged": "^13.2.2",
+        "husky": "^9.1.7",
+        "jest": "^30.0.4",
+        "lint-staged": "^16.1.2",
         "openapi-types": "^12.1.0",
-        "prettier": "2.8.1",
-        "prettier-eslint": "15.0.1",
-        "rimraf": "^5.0.1",
+        "prettier": "3.6.2",
+        "prettier-eslint": "16.4.2",
+        "rimraf": "^6.0.1",
         "ts-jest": "^29.1.0",
-        "typescript": "^5.0.4",
-        "vsce": "^2.15.0"
+        "@vscode/vsce": "^3.6.0"
     },
+    "type": "module",
     "dependencies": {
-        "@apidevtools/swagger-parser": "^10.1.0",
-        "express": "^4.18.2",
+        "@apidevtools/swagger-parser": "^12.0.0",
+        "express": "^5.1.0",
         "lodash": "^4.17.21",
         "portfinder": "^1.0.32",
         "socket.io": "^4.6.1",
         "socket.io-client": "^4.6.1",
-        "swagger-ui-dist": "^4.18.3"
+        "swagger-ui-dist": "^5.26.0"
     },
     "bugs": {
         "url": "https://github.com/v-swagger/v-swagger/issues"
